# contracts/strategy-signal-service.openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "strategy-signal-service",
    "description" : "Strategy registry + live signal engine (Stage C): D18 lifecycle, signals with score breakdowns",
    "version" : "v1"
  },
  "paths" : {
...
    "/api/v1/paper/account" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "account",
        "parameters" : [ {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountDto"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "updateAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountDto"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/strategies/validate" : {
      "post" : {
        "tags" : [ "registry-controller" ],
        "operationId" : "validate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidateResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/paper/reset" : {
      "post" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "reset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetBody"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/positions/{id}/close" : {
      "post" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "close",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CloseBody"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TradeDto"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/orders" : {
      "post" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "order",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrderBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PositionDto"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/paper/positions/{id}/brackets" : {
      "patch" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "editBrackets",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BracketBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PositionDetail"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/paper/trades" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "trades",
        "parameters" : [ {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "symbol",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TradePage"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/trades/export" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "tradesExport",
        "parameters" : [ {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "symbol",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/positions" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "positions",
        "parameters" : [ {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PositionList"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/positions/{id}" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "positionDetail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PositionDetail"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/pnl" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "pnl",
        "parameters" : [ {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pnl"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/paper/events" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "events",
        "parameters" : [ {
          "name" : "positionId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "day",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaperEventsResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paper/attribution" : {
      "get" : {
        "tags" : [ "paper-controller" ],
        "operationId" : "attribution",
        "parameters" : [ {
          "name" : "book",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Attribution"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/positions" : {
      "get" : {
        "tags" : [ "orders-controller" ],
        "operationId" : "positions_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Positions"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AccountBody" : {
        "type" : "object",
        "properties" : {
          "startingCapital" : {
            "type" : "number"
          },
          "book" : {
            "type" : "string"
          }
        }
      },
      "AccountDto" : {
        "type" : "object",
        "properties" : {
          "startingCapital" : {
            "type" : "string"
          },
          "cash" : {
            "type" : "string"
          },
          "equity" : {
            "type" : "string"
          },
          "realized" : {
            "type" : "string"
          },
          "unrealized" : {
            "type" : "string"
          },
          "dayPnl" : {
            "type" : "string"
          },
          "openPositions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "capitalUsed" : {
            "type" : "string"
          },
          "usageByClass" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "marginPercents" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "capitalUsed", "cash", "dayPnl", "equity", "marginPercents", "openPositions", "realized", "startingCapital", "unrealized", "usageByClass" ]
      },
...
      "ValidateRequest" : {
        "type" : "object",
        "properties" : {
          "config" : {
            "type" : "string"
          }
        }
      },
...
      "ResetBody" : {
        "type" : "object",
        "properties" : {
          "confirm" : {
            "type" : "boolean"
          },
          "book" : {
            "type" : "string"
          }
        }
      },
      "CloseBody" : {
        "type" : "object",
        "properties" : {
          "price" : {
            "type" : "number"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object"
          }
        }
      }
    }
  }
}